Move Synonyms to server
