+++ server/src/main/java/org/elasticsearch/synonyms/SynonymSet.java
@@ -6,7 +6,7 @@
  * Side Public License, v 1.
  */
 
-package org.elasticsearch.analysis.common.synonyms;
+package org.elasticsearch.synonyms;
 
 import org.elasticsearch.common.io.stream.StreamInput;
 import org.elasticsearch.common.io.stream.StreamOutput;
